changes in weatherman
